requirements.txt
numpy
pyyaml
matplotlib
scipy
pandas
setuptools
tensorflow_gpu
